+++ Modules/_testcapi/pyatomic.c
@@ -8,7 +8,7 @@
 #undef NDEBUG
 
 #include "Python.h"
-#include "pyatomic.h"
+#include "cpython/pyatomic.h"
 #include "parts.h"
 
 // We define atomic bitwise operations on these types
